updated fetch function
